requirements.txt
apache-airflow[celery]==3.1.2
pandas
pydantic>=2.0
pydantic-settings>=2.9.1